pyproject.toml
[tool.poetry]
name = "aiogram-test"
version = "0.1.0"
description = ""
authors = ["Borteq2 <borteq2@yandex.ru>"]
readme = "README.md"
packages = [{include = "aiogram_test"}]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^2.25.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
